super final commit

+++ src/main/java/com/deynak/dao/UserDaoHibernateImpl.java
@@ -4,6 +4,7 @@ import com.deynak.model.User;
 import com.deynak.util.Util;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
+import org.hibernate.Transaction;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,55 +16,64 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
-        try (Session session = Util.getConnectionHibernate().openSession();) {
-            session.beginTransaction();
+        Transaction transaction = null;
+        try (Session session = Util.getConnectionHibernate().openSession()) {
+            transaction = session.beginTransaction();
             String hql = "CREATE TABLE IF NOT EXISTS hiber_User (id BIGINT PRIMARY KEY AUTO_INCREMENT, name VARCHAR(50), lastName VARCHAR(50), age INT)";
             session.createNativeQuery(hql).executeUpdate();
-            session.getTransaction().commit();
+            transaction.commit();
         } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
-
     }
 
     @Override
     public void dropUsersTable() {
-        try (Session session = Util.getConnectionHibernate().openSession();) {
-            session.beginTransaction();
+        Transaction transaction = null;
+        try (Session session = Util.getConnectionHibernate().openSession()) {
+            transaction = session.beginTransaction();
             session.createNativeQuery("DROP TABLE IF EXISTS hiber_User").executeUpdate();
-            session.getTransaction().commit();
+            transaction.commit();
         } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
-        try (Session session = Util.getConnectionHibernate().openSession();) {
-            session.beginTransaction();
+        Transaction transaction = null;
+        try (Session session = Util.getConnectionHibernate().openSession()) {
+            transaction = session.beginTransaction();
             session.save(new User(name, lastName, age));
-            session.getTransaction().commit();
+            transaction.commit();
         } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
     }
 
     @Override
     public void removeUserById(long id) {
-        try (Session session = Util.getConnectionHibernate().openSession();) {
-            session.beginTransaction();
+        Transaction transaction = null;
+        try (Session session = Util.getConnectionHibernate().openSession()) {
+            transaction = session.beginTransaction();
             User user = session.get(User.class, id);
-            session.delete(user);
+            if (user != null) {
+                session.delete(user);
+                transaction.commit();
+            }
         } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
-
     }
 
     @Override
     public List<User> getAllUsers() {
         List<User> userList = new ArrayList<>();
-        try (Session session = Util.getConnectionHibernate() .openSession()) {
+        try (Session session = Util.getConnectionHibernate().openSession()) {
             userList = session.createQuery("from User", User.class).list();
         } catch (HibernateException e) {
             e.printStackTrace();
@@ -74,11 +84,13 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void cleanUsersTable() {
-        try (Session session = Util.getConnectionHibernate().openSession();) {
-            session.beginTransaction();
+        Transaction transaction = null;
+        try (Session session = Util.getConnectionHibernate().openSession()) {
+            transaction = session.beginTransaction();
             session.createQuery("DELETE FROM User").executeUpdate();
-            session.getTransaction().commit();
+            transaction.commit();
         } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
     }
